# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nurul Quran API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface Speaker {
  id: number;
  name: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  lectureCount: number;
}

export interface Lecture {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  audioUrl: string;
  /** @nullable */
  duration?: number | null;
  language: string;
  category: string;
  isPremium: boolean;
  isFeatured: boolean;
  playCount: number;
  /** @nullable */
  thumbnailUrl?: string | null;
  /** @nullable */
  speakerId?: number | null;
  /** @nullable */
  speakerName?: string | null;
  createdAt: string;
}

export interface LectureListResponse {
  lectures: Lecture[];
  total: number;
  limit: number;
  offset: number;
}

export interface Course {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  category: string;
  isPremium: boolean;
  lectureCount: number;
  /** @nullable */
  thumbnailUrl?: string | null;
  /** @nullable */
  speakerId?: number | null;
  /** @nullable */
  speakerName?: string | null;
  createdAt: string;
}

export interface DailyAyah {
  id: number;
  arabicText: string;
  translation: string;
  surahName: string;
  surahNumber: number;
  ayahNumber: number;
  reference: string;
}

export type DashboardSummaryCategoriesItem = {
  name: string;
  count: number;
};

export interface DashboardSummary {
  totalLectures: number;
  totalCourses: number;
  totalSpeakers: number;
  totalListeners: number;
  categories: DashboardSummaryCategoriesItem[];
}

export interface User {
  id: number;
  email: string;
  name: string;
  isPremium: boolean;
  createdAt: string;
}

export interface RegisterUserBody {
  email: string;
  /** @minLength 8 */
  password: string;
  name: string;
}

export interface LoginUserBody {
  email: string;
  password: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface CreateCheckoutSessionBody {
  /** Amount in cents */
  amount: number;
  currency?: string;
  successUrl: string;
  cancelUrl: string;
}

export interface CreateSubscriptionBody {
  successUrl: string;
  cancelUrl: string;
  /** @nullable */
  userId?: number | null;
}

export interface CheckoutSessionResponse {
  url: string;
  sessionId: string;
}

export type GetLecturesParams = {
  search?: string;
  /**
   * @nullable
   */
  speakerId?: number | null;
  /**
   * @nullable
   */
  language?: GetLecturesLanguage;
  /**
   * @nullable
   */
  category?: string | null;
  /**
   * @nullable
   */
  isPremium?: boolean | null;
  limit?: number;
  offset?: number;
};

export type GetLecturesLanguage =
  | (typeof GetLecturesLanguage)[keyof typeof GetLecturesLanguage]
  | null;

export const GetLecturesLanguage = {
  English: "English",
  Urdu: "Urdu",
} as const;

export type GetRecentLecturesParams = {
  limit?: number;
};

export type GetCoursesParams = {
  /**
   * @nullable
   */
  category?: string | null;
  /**
   * @nullable
   */
  isPremium?: boolean | null;
};

export type GetRecentlyPlayedParams = {
  limit?: number;
};