# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Nurul Quran API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /lectures:
    get:
      operationId: getLectures
      tags: [lectures]
      summary: List all lectures
      description: Returns lectures with optional filters
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: speakerId
          in: query
          schema:
            type: ["integer", "null"]
        - name: language
          in: query
          schema:
            type: ["string", "null"]
            enum: ["English", "Urdu", null]
        - name: category
          in: query
          schema:
            type: ["string", "null"]
        - name: isPremium
          in: query
          schema:
            type: ["boolean", "null"]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of lectures
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LectureListResponse"
  /lectures/recent:
    get:
      operationId: getRecentLectures
      tags: [lectures]
      summary: Get recent lectures
      description: Returns the most recently added lectures
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 6
      responses:
        "200":
          description: Recent lectures
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lecture"
...
  /lectures/{id}:
    get:
      operationId: getLecture
      tags: [lectures]
      summary: Get a single lecture
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lecture detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lecture"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /courses:
    get:
      operationId: getCourses
      tags: [courses]
      summary: List all courses
      description: Returns all courses with optional filters
      parameters:
        - name: category
          in: query
          schema:
            type: ["string", "null"]
        - name: isPremium
          in: query
          schema:
            type: ["boolean", "null"]
      responses:
        "200":
          description: List of courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"
  /courses/{id}:
    get:
      operationId: getCourse
      tags: [courses]
      summary: Get a single course
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Course detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /payments/create-checkout-session:
    post:
      operationId: createCheckoutSession
      tags: [payments]
      summary: Create a Stripe checkout session for donation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCheckoutSessionBody"
      responses:
        "200":
          description: Checkout session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutSessionResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /payments/create-subscription:
    post:
      operationId: createSubscription
      tags: [payments]
      summary: Create a Stripe subscription session for premium
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubscriptionBody"
      responses:
        "200":
          description: Subscription session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutSessionResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Lecture:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        audioUrl:
          type: string
        duration:
          type: ["integer", "null"]
        language:
          type: string
        category:
          type: string
        isPremium:
          type: boolean
        isFeatured:
          type: boolean
        playCount:
          type: integer
        thumbnailUrl:
          type: ["string", "null"]
        speakerId:
          type: ["integer", "null"]
        speakerName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - audioUrl
        - language
        - category
        - isPremium
        - isFeatured
        - playCount
        - createdAt
    LectureListResponse:
      type: object
      properties:
        lectures:
          type: array
          items:
            $ref: "#/components/schemas/Lecture"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
      required:
        - lectures
        - total
        - limit
        - offset
    Course:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
        isPremium:
          type: boolean
        lectureCount:
          type: integer
        thumbnailUrl:
          type: ["string", "null"]
        speakerId:
          type: ["integer", "null"]
        speakerName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - category
        - isPremium
        - lectureCount
        - createdAt
...
    CreateCheckoutSessionBody:
      type: object
      properties:
        amount:
          type: integer
          description: Amount in cents
        currency:
          type: string
          default: usd
        successUrl:
          type: string
        cancelUrl:
          type: string
      required:
        - amount
        - successUrl
        - cancelUrl
    CreateSubscriptionBody:
      type: object
      properties:
        successUrl:
          type: string
        cancelUrl:
          type: string
        userId:
          type: ["integer", "null"]
      required:
        - successUrl
        - cancelUrl
    CheckoutSessionResponse:
      type: object
      properties:
        url:
          type: string
        sessionId:
          type: string
      required:
        - url
        - sessionId