# lib/api-zod/src/generated/types/course.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nurul Quran API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Course {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  category: string;
  isPremium: boolean;
  lectureCount: number;
  /** @nullable */
  thumbnailUrl?: string | null;
  /** @nullable */
  speakerId?: number | null;
  /** @nullable */
  speakerName?: string | null;
  createdAt: Date;
}