# lib/api-zod/src/generated/types/createCheckoutSessionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nurul Quran API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateCheckoutSessionBody {
  /** Amount in cents */
  amount: number;
  currency?: string;
  successUrl: string;
  cancelUrl: string;
}