# lib/api-zod/src/generated/types/lecture.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nurul Quran API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Lecture {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  audioUrl: string;
  /** @nullable */
  duration?: number | null;
  language: string;
  category: string;
  isPremium: boolean;
  isFeatured: boolean;
  playCount: number;
  /** @nullable */
  thumbnailUrl?: string | null;
  /** @nullable */
  speakerId?: number | null;
  /** @nullable */
  speakerName?: string | null;
  createdAt: Date;
}